backend/requirements.txt
asgiref==3.7.2
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-filter==23.5
django-rest-passwordreset==1.4.0
djangorestframework==3.14.0
gunicorn==22.0.0
idna==3.7
packaging==24.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.4.4
stripe==8.10.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
